Revert "upgrade to 1.5.5.0"
