+++ src/compat.h
@@ -25,13 +25,9 @@
 #include <net/if.h>
 #include <netinet/in.h>
 #include <ifaddrs.h>
+#endif
 
-#ifndef _MSC_VER
 typedef u_int SOCKET;
-#endif
-#endif
-
-
 #ifdef WIN32
 #define MSG_NOSIGNAL        0
 #define MSG_DONTWAIT        0
