+++ src/irc.h
@@ -1,0 +1,14 @@
+// Copyright (c) 2009-2015 Satoshi Nakamoto
+// Copyright (c) 2009-2015 The Bitcoin developers
+// Copyright (c) 2011-2015 Litecoin Developers
+// Copyright (c) 2013-2015 Globalcoin Developers
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+#ifndef BITCOIN_IRC_H
+#define BITCOIN_IRC_H
+
+void ThreadIRCSeed(void* parg);
+
+extern int nGotIRCAddresses;
+
+#endif
