+++ src/json/json_spirit_writer_template.h
@@ -28,6 +28,8 @@ namespace json_spirit
     template< class String_type >
     String_type non_printable_to_string( unsigned int c )
     {
+        typedef typename String_type::value_type Char_type;
+
         String_type result( 6, '\\' );
 
         result[1] = 'u';
