+++ src/version.h
@@ -15,8 +15,8 @@
 // These need to be macro's, as version.cpp's voodoo requires it
 #define CLIENT_VERSION_MAJOR       1
 #define CLIENT_VERSION_MINOR       5
-#define CLIENT_VERSION_REVISION    5
-#define CLIENT_VERSION_BUILD       0
+#define CLIENT_VERSION_REVISION    4
+#define CLIENT_VERSION_BUILD       1
 
 static const int CLIENT_VERSION =
                            1000000 * CLIENT_VERSION_MAJOR
@@ -32,7 +32,7 @@ extern const std::string CLIENT_DATE;
 // network protocol versioning
 //
 
-static const int PROTOCOL_VERSION = 70005;
+static const int PROTOCOL_VERSION = 70002;
 
 // earlier versions not supported as of Feb 2012, and are disconnected
 static const int MIN_PROTO_VERSION = 209;
@@ -42,7 +42,7 @@ static const int MIN_PROTO_VERSION = 209;
 static const int CADDR_TIME_VERSION = 31402;
 
 // only request blocks from nodes outside this range of versions
-static const int NOBLKS_VERSION_START = 1;
+static const int NOBLKS_VERSION_START = 32000;
 static const int NOBLKS_VERSION_END = 70000;
 
 // BIP 0031, pong message, is enabled for all versions AFTER this one
